# spring_ajax_jpa/openapi.yaml
openapi: 3.0.1
info:
  title: Game Management Service
  description: API for managing computer games, developers and publishers
  version: 1.0.0
servers:
  - url: http://localhost:8080/ComputerGames
...
paths:
...
  /publishers:
    get:
      tags: [Publishers]
      summary: Get all publishers
      responses:
        '200':
          description: List of all publishers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publisher'
        
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '403':
          description: Authentication granted but user does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
          
    post:
      tags: [Publishers]
      summary: Create a new publisher
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Game publisher name
                  example: CD Projekt
                country:
                  type: string
                  description: Game publisher country
                  example: Poland
              required:
                - name
                - country
                
      responses:
        '201':
          description: Publisher created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
              
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '403':
          description: Authentication granted but user does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
  /publishers/{id}:
    delete:
      tags: [Publishers]
      summary: Delete a publisher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Publisher deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '403':
          description: Authentication granted but user does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '404':
          description: Publisher not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
    
    get:
      tags: [Publishers]
      summary: Get a publisher by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Publisher finded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publisher'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
          
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '403':
          description: Authentication granted but user does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '404':
          description: Publisher not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
components:
  schemas:
...
    Publisher:
      type: object
      properties:
        id:
          type: integer
          description: Game publisher identifier
          example: 1
        name:
          type: string
          description: Game publisher name
          example: CD Projekt
        country:
          type: string
          description: Game publisher county
          example: Poland
      required:
        - id
        - name
        - country
    OperationStatus:
      type: object
      properties:
        status:
          type: string
          description: Description of the operation execution status
          example: Operation error, invalid fields